refactor: TAP-5862 Upgrade jdk8 to jdk17
Closes TAP-5862

+++ hazelcast-persistence/src/main/java/com/hazelcast/persistence/store/impl/MongoDBIMap.java
@@ -65,7 +65,7 @@ public class MongoDBIMap extends PersistenceMapStore<PersistenceMongoDBConfig, M
 
 	@Override
 	public void destroy() {
-		if (this.persistenceMongoDBConfig.isExclusiveCollection()) {
+		if (this.persistenceMongoDBConfig.isExclusiveCollection() && null != mongoDBResource) {
 			mongoDBResource.getMongoCollection().drop();
 		} else {
 			this.deleteAll(null);
